% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{calcDivisions}
\alias{calcDivisions}
\title{Calculate histogram binning by quantile.}
\usage{
calcDivisions(vec, q, bins = NULL, transformed = FALSE)
}
\arguments{
\item{vec}{A vector of numbers.}

\item{q}{The quantile to use for cutoffs.}

\item{bins}{The number of bins to use for a histogram.}

\item{transformed}{Adjust divisions if transformed}
}
\description{
Modified from: GenomicDistributions (Tessa Danehy)
}
\examples{
calcDivisions()
}
\keyword{quantiles}
