% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{calcQuantileCutoff}
\alias{calcQuantileCutoff}
\title{Determine which quantile to use as cutoff.}
\usage{
calcQuantileCutoff(vec, baseline = 1)
}
\arguments{
\item{vec}{A vector of numbers.}

\item{baseline}{A minimum quantile cutoff.}
}
\description{
Modified from: GenomicDistributions (Tessa Danehy)
}
\examples{
calcQuantileCutoff()
}
\keyword{cutoff}
\keyword{quantiles}
