% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{cutDists}
\alias{cutDists}
\title{Create a character vector of bin labels.}
\usage{
cutDists(
  vec,
  divisions = c(-Inf, -1e+06, -10000, -1000, -100, 0, 100, 1000, 10000, 1e+06, Inf)
)
}
\arguments{
\item{vec}{A vector (or list of vectors) of numbers.}

\item{divisions}{The break points of the distribution}
}
\description{
Modified from: GenomicDistributions (Tessa Danehy)
}
\examples{
cutDists()
}
\keyword{quantiles}
