% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{fancyNumbers}
\alias{fancyNumbers}
\title{Custom formatting function for the log axis.}
\usage{
fancyNumbers(n)
}
\arguments{
\item{n}{A vector of numbers.}
}
\description{
From:
DeCicco, L. (2018, August 10). Exploring ggplot2 boxplots - 
 Defining limits and adjusting style [Blog post]. 
 Retrieved from https://waterdata.usgs.gov/blog/boxplots/
}
\examples{
fancyNumbers()
}
\keyword{limits}
