% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{labelCuts}
\alias{labelCuts}
\title{Create a character vector of bin labels.}
\usage{
labelCuts(breakPoints, digits = 1, collapse = "-", infBins = FALSE)
}
\arguments{
\item{breakPoints}{A vector of numbers.}

\item{digits}{The number of digits to round to.}

\item{collapse}{The character to separate paste values by.}

\item{infBins}{Whether to use infinite bins.}
}
\description{
From: GenomicDistributions (Tessa Danehy)
}
\examples{
labelCuts()
}
\keyword{labels}
