% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{n_fun}
\alias{n_fun}
\title{Return the count of the plot at 95% of the upper limit.}
\usage{
n_fun(x, lim = 50)
}
\arguments{
\item{x}{A vector of numbers.}

\item{lim}{Upper limit.}
}
\description{
From:
DeCicco, L. (2018, August 10). Exploring ggplot2 boxplots - 
 Defining limits and adjusting style [Blog post]. 
 Retrieved from https://waterdata.usgs.gov/blog/boxplots/
}
\examples{
n_fun()
}
\keyword{limits}
