% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{plotAdapt}
\alias{plotAdapt}
\title{Plot the distribution of adapter insertions}
\usage{
plotAdapt(input, name = "adapt", umi_len = 0)
}
\arguments{
\item{input}{FLASH histogram output}

\item{name}{A sample name or identifier for the plot title}

\item{umi_len}{The UMI length}
}
\description{
Plot the distribution of adapter insertions
}
\examples{
data("adapt")
plotAdapt(input = "adapt")
}
\keyword{cutadapt}
