% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{plotCutadapt}
\alias{plotCutadapt}
\title{Plot the cutadapt-based distribution of adapter insertions}
\usage{
plotCutadapt(input, name = "cutadapt", umi_len = 0, count_factor = 1e+06)
}
\arguments{
\item{input}{A cutadapt report}

\item{name}{A sample name or identifier for the plot title}

\item{umi_len}{The UMI length}
}
\description{
Plot the cutadapt-based distribution of adapter insertions
}
\examples{
data("cutadapt")
plotCutadapt(input = "cutadapt")
}
\keyword{cutadapt}
