% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{plotPI}
\alias{plotPI}
\title{Plot the distribution of highest covered TSS density/gene body density ratios}
\usage{
plotPI(
  pi,
  name = "pause indicies",
  raw = FALSE,
  type = c("histogram", "boxplot", "violin"),
  annotate = TRUE
)
}
\arguments{
\item{pi}{A single column containing the ratio of TSS densities/gene body
densities for the highest scoring TSSs}

\item{name}{X-axis label, typically a sample name}

\item{type}{Plot format}

\item{annotate}{Display mean and median values on plot}
}
\description{
Plot the distribution of highest covered TSS density/gene body density ratios
}
\examples{
data("pidx")
plotPI(pi = "pidx")
}
\keyword{index}
\keyword{pause}
