% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{prettyLogs}
\alias{prettyLogs}
\title{This function forces the y-axis breaks to be on every 10^x.}
\usage{
prettyLogs(x)
}
\arguments{
\item{x}{A vector of numbers.}
}
\description{
From:
DeCicco, L. (2018, August 10). Exploring ggplot2 boxplots - 
 Defining limits and adjusting style [Blog post]. 
 Retrieved from https://waterdata.usgs.gov/blog/boxplots/
}
\examples{
prettyLogs()
}
\keyword{limits}
