% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{theme_PEPPRO}
\alias{theme_PEPPRO}
\title{A standardized ggplot theme for PEPPRO plots}
\usage{
theme_PEPPRO(base_family = "sans", ...)
}
\description{
A standardized ggplot theme for PEPPRO plots
}
\examples{
theme_PEPPRO()
}
\keyword{ggplot2}
\keyword{theme}
