% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEPPROr.R
\name{readPepproGeneCounts}
\alias{readPepproGeneCounts}
\title{Read PEPPRO gene coverage files}
\usage{
readPepproGeneCounts(project, results_subdir)
}
\arguments{
\item{results_subdir}{A PEP project results subdirectory path}

\item{pep}{A PEPr project object}
}
\description{
Read PEPPRO gene coverage files
}
\keyword{coverage}
\keyword{gene}
\keyword{project}
